Compile Time Execution initial implementation. Procs IR and local values have been moved out of fiber data and into the Procedure. String literal structs are no longer written to static data segment (the string data itself still is, of course). Oh, and added SourceLocations to every IRInstruction (might have missed some).

+++ src/Strings.cpp
@@ -1,11 +1,11 @@
 inline String operator""_s(const char *str, u64 size)
 {
-	return { (s64)size, str };
+	return { size, str };
 }
 
 inline String TStringConcat(String a, String b)
 {
-	s64 size = a.size + b.size;
+	u64 size = a.size + b.size;
 	char *buffer = (char *)ThreadAllocator::Alloc(size, 1);
 	String result = { size, buffer };
 	memcpy(buffer, a.data, a.size);
@@ -15,7 +15,7 @@ inline String TStringConcat(String a, String b)
 
 inline String SStringConcat(String a, String b)
 {
-	s64 size = a.size + b.size;
+	u64 size = a.size + b.size;
 	char *buffer = (char *)LinearAllocator::Alloc(size, 1);
 	String result = { size, buffer };
 	memcpy(buffer, a.data, a.size);
@@ -33,7 +33,7 @@ const char *StringToCStr(String str, void *(*allocFunc)(u64, int))
 
 String CStrToString(const char *cstr)
 {
-	return { (s64)strlen(cstr), cstr };
+	return { strlen(cstr), cstr };
 }
 
 inline void ChangeExtension(char *buffer, const char *newExtension)
